Adjust method model
