+++ stock_lot_suffix/models/__init__.py
@@ -2,4 +2,4 @@ from . import account_analytic_account
 from . import purchase_order
 from . import res_partner
 from . import stock_lot
-from . import stock_picking
+from . import stock_move_line
