+++ stock_lot_suffix/models/stock_lot.py
@@ -8,19 +8,17 @@ class StockLot(models.Model):
     _inherit = "stock.lot"
 
     channel_category = fields.Char()
-    analytic_category = fields.Char()
+    lot_suffix = fields.Char()
 
     def name_get(self):
         name_list = super(StockLot, self).name_get()
         new_name_list = []
         for name in name_list:
             rec = self.browse(name[0])
-            channel = rec.channel_category
-            analytic = rec.analytic_category
             new_name = name[1]
-            if analytic:
-                new_name += analytic
-            if channel:
-                new_name += channel
+            if rec.lot_suffix:
+                new_name += rec.lot_suffix
+            if rec.channel_category:
+                new_name += rec.channel_category
             new_name_list.append((name[0], new_name))
         return new_name_list
