+++ stock_lot_suffix/views/account_analytic_account_views.xml
@@ -6,7 +6,7 @@
         <field name="inherit_id" ref="analytic.view_account_analytic_account_form" />
         <field name="arch" type="xml">
             <xpath expr="//field[@name='code']" position="after">
-                <field name="analytic_category" />
+                <field name="lot_suffix" />
             </xpath>
         </field>
     </record>
